local_requirements.txt
diffusers==0.27.2
torch==2.1.0
pillow==10.3.0
boto3
botocore